Added some SHA and AES tests for simulation, fixed a icarus specific bug
