+++ software/sw_build.mk
@@ -78,10 +78,16 @@ IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/iob_soc_opencryptolinux_firmware.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/printf.c
 
 # NOTE(Ruben): To speed up simulation, we do not include or simulate crypto code in simulation. It greatly increases binary size and some tests would take forever. Better to run all tests in fpga-run.
-ifneq ($(SIMULATION),1)
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto.c
-IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto_tests.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/crypto/aes.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto_common_tests.c
+ifeq ($(SIMULATION),1)
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_simple_crypto_tests.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/crypto/McEliece/arena.c)
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/crypto/McEliece/common/sha2.c)
+else
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_crypto_tests.c
+IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=src/versat_mceliece.c
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/crypto/McEliece/*.c)
 IOB_SOC_OPENCRYPTOLINUX_FW_SRC+=$(wildcard src/crypto/McEliece/common/*.c)
 endif
